+++ Packages/FluentParticleSystem/Runtime/MainModuleExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2710ff29301a464cbf61008380606c4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
