+++ Packages/FluentParticleSystem/Tests/Runtime/MainModuleExtensionTest.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace OUCC.FluentParticleSystem.Test
+{
+    public class MainModuleExtensionTest
+    {
+        public void Test() {
+            var main = new GameObject().AddComponent<ParticleSystem>().main;
+
+        }
+    }
+}
