+++ Packages/FluentParticleSystem/Tests/Runtime/MainModuleExtensionTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f4923bfcf2ec46f4e8ca0d7ad182d58d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
