+++ Packages/FluentParticleSystem/Tests/Runtime/Unity.FluentParticleSystem.Test.asmdef
@@ -1,6 +1,6 @@
 {
     "name": "Unity.FluentParticleSystem.Test",
-    "rootNamespace": "FluentParticleSystem.Test",
+    "rootNamespace": "OUCC.FluentParticleSystem.Test",
     "references": [
         "GUID:27619889b8ba8c24980f49ee34dbb44a",
         "GUID:0acc523941302664db1f4e527237feb3",
